auth_system/requirements.txt
aioredis==2.0.1
aiosmtplib==3.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
black==24.3.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fonttools==4.51.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openai==1.30.1
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydentic==0.0.1.dev3
pydyf==0.10.0
PyJWT==2.8.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
pytils==0.4.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
stripe==9.3.0
tinycss2==1.3.0
tomli==2.0.1
tools==0.1.9
tqdm==4.66.4
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
zopfli==0.2.3
